ref(ricer_test_tools)!: Refactor test stubs into test fixtures
BREAKING CHANGES: Renamed test stub tools into test fixtures, because all they
do is define data fixtures _not_ stubs. Plus, made their interfaces much more
friendlier to use.

TODO: Still need to add actual Git functionality to the Git fixture object.
