+++ crates/ricer_test_tools/src/lib.rs
@@ -2,5 +2,4 @@
 // SPDX-License-Identifier: GPL-2.0-or-later WITH GPL-CC-1.0
 
 pub mod fakes;
-
-pub mod stubs;
+pub mod fixtures;
